version 3.0.0

+++ source/Apm4D.Share.Response.pas
@@ -16,14 +16,14 @@ type
   TResponse = class
   private
     FStatus_code: Integer;
-    Fencoded_body_size: double;
-    Ftransfer_size: double;
-    Fdecoded_body_size: double;
+    Fencoded_body_size: Int64;
+    Ftransfer_size: Int64;
+    Fdecoded_body_size: Int64;
   public
     constructor Create(const AStatusCode: Integer);
-    property decoded_body_size: double read Fdecoded_body_size write Fdecoded_body_size;
-    property encoded_body_size: double read Fencoded_body_size write Fencoded_body_size;
-    property transfer_size: double read Ftransfer_size write Ftransfer_size;
+    property decoded_body_size: Int64 read Fdecoded_body_size write Fdecoded_body_size;
+    property encoded_body_size: Int64 read Fencoded_body_size write Fencoded_body_size;
+    property transfer_size: Int64 read Ftransfer_size write Ftransfer_size;
   end;
 
 implementation
